Close InnRoom chest immediately upon moving
Right now, the InnRoom chest is a hidden container in the player's bank boxes that is moved to their feet and opened.

As with other regular containers, moving while the container is open does not close them until the player gets out of range.

However, as most players treat the InnRoom chest as an extension of their bank, this behavior is not necessary and can often lead to losing the chest.

This is due to the fact they can enter the inn, open their chest while near the exit door, then leave without taking enough steps away to "close" and return the chest to their bank.

This change closes the InnRoom chest immediately when the player moves, much like how the bank box works, to prevent these instances of loss.

It requires a World.exe rebuild.